Always unset elligator bit (ECVRF v3)
In order to be compatible with the version 3 of the VRF standard, we need to always
unset the sign bit in the elligator function. Given the exposure of functions in
this library, we need to fork it. This, unfortunately, is required for legacy code.

Only changes, with respect to version 3.2.0, are the sign bit and the corresponding
tests.